### 2023-03-13 #5252 @JorenSix
Another potential reviewer: [hendriks73](https://github.com/hendriks73) During [my last JOSS paper](https://github.com/openjournals/joss-reviews/issues/4554) I learned a thanks to his feedback and involvement.
