+++ search/all_a.js
@@ -17,74 +17,88 @@ var searchData=
   ['olaf_5fdb_5ffind_5fmeta_5fdata_14',['olaf_db_find_meta_data',['../olaf__db_8h.html#ae71544a8e313df918fb6ad5d97a82d02',1,'olaf_db.c']]],
   ['olaf_5fdb_5ffind_5fsingle_15',['olaf_db_find_single',['../olaf__db_8h.html#ac0c36d0f2fa63d8477e5ea96aa7439f2',1,'olaf_db.c']]],
   ['olaf_5fdb_5fhas_5fmeta_5fdata_16',['olaf_db_has_meta_data',['../olaf__db_8h.html#ab4044ad27aa8eb322e186c5b130edc99',1,'olaf_db.c']]],
-  ['olaf_5fdb_5fnew_17',['olaf_db_new',['../olaf__db_8h.html#aeac2be6db9d807b888aee672ba707f56',1,'olaf_db.c']]],
-  ['olaf_5fdb_5fstats_18',['olaf_db_stats',['../olaf__db_8h.html#a0eeebdfb817fa14eda52370dc77daf6f',1,'olaf_db.c']]],
-  ['olaf_5fdb_5fstats_5fmeta_5fdata_19',['olaf_db_stats_meta_data',['../olaf__db_8h.html#a0ba7f001338a3df72b495229f6579dee',1,'olaf_db.c']]],
-  ['olaf_5fdb_5fstore_20',['olaf_db_store',['../olaf__db_8h.html#a61c00f7588f977094007126da8699860',1,'olaf_db.c']]],
-  ['olaf_5fdb_5fstore_5fmeta_5fdata_21',['olaf_db_store_meta_data',['../olaf__db_8h.html#a96a6d1f7e683764fb24863560c6ea71f',1,'olaf_db.c']]],
-  ['olaf_5fdb_5fstring_5fhash_22',['olaf_db_string_hash',['../olaf__db_8h.html#af31efb5f40fe0ec0f0c2db0f77a31a23',1,'olaf_db.c']]],
-  ['olaf_5fdeque_23',['Olaf_Deque',['../struct_olaf___deque.html',1,'']]],
-  ['olaf_5fdeque_2eh_24',['olaf_deque.h',['../olaf__deque_8h.html',1,'']]],
-  ['olaf_5fep_5fextractor_25',['Olaf_EP_Extractor',['../struct_olaf___e_p___extractor.html',1,'']]],
-  ['olaf_5fep_5fextractor_2eh_26',['olaf_ep_extractor.h',['../olaf__ep__extractor_8h.html',1,'']]],
-  ['olaf_5fep_5fextractor_5fdestroy_27',['olaf_ep_extractor_destroy',['../olaf__ep__extractor_8h.html#a69adabca5309ae71408f42a28f71e2c6',1,'olaf_ep_extractor.c']]],
-  ['olaf_5fep_5fextractor_5fextract_28',['olaf_ep_extractor_extract',['../olaf__ep__extractor_8h.html#a107ed8df3d2e7f3203f63af7bbeccbf2',1,'olaf_ep_extractor.c']]],
-  ['olaf_5fep_5fextractor_5fnew_29',['olaf_ep_extractor_new',['../olaf__ep__extractor_8h.html#a3e570ee9bda544ac3141eaa84e46b54a',1,'olaf_ep_extractor.c']]],
-  ['olaf_5fep_5fextractor_5fprint_5fep_30',['olaf_ep_extractor_print_ep',['../olaf__ep__extractor_8h.html#a2247360eca5e559b344925e2a8b44289',1,'olaf_ep_extractor.c']]],
[…]
