+++ search/variables_a.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['searchrange_0',['searchRange',['../struct_olaf___config.html#af06994a7fd41623b206781b4aa973548',1,'Olaf_Config']]]
+  ['path_0',['path',['../struct_olaf___resource___meta__data.html#a06f9246831bebc65cba89365de74b17e',1,'Olaf_Resource_Meta_data']]],
+  ['printresultevery_1',['printResultEvery',['../struct_olaf___config.html#a969a9165d5ad8085b7ace679e4ac7221',1,'Olaf_Config']]]
 ];
