+++ search/searchdata.js
@@ -1,10 +1,10 @@
 var indexSectionsWithContent =
 {
-  0: "_abdefhkmnopqsv",
+  0: "_abdefhkmnopqsuv",
   1: "_efmop",
   2: "hopq",
   3: "hopq",
-  4: "abdefhkmnpsv",
+  4: "abdefhkmnopsuv",
   5: "hq",
   6: "p",
   7: "hq",
