+++ search/variables_9.js
@@ -1,5 +1,8 @@
 var searchData=
 [
-  ['path_0',['path',['../struct_olaf___resource___meta__data.html#a06f9246831bebc65cba89365de74b17e',1,'Olaf_Resource_Meta_data']]],
-  ['printresultevery_1',['printResultEvery',['../struct_olaf___config.html#a969a9165d5ad8085b7ace679e4ac7221',1,'Olaf_Config']]]
+  ['olaf_5fdb_5fmem_5faudio_5fduration_0',['olaf_db_mem_audio_duration',['../olaf__fp__ref__mem_8h.html#a4ab864697ab6bcf7ef244addb9d73eb3',1,'olaf_fp_ref_mem.h']]],
+  ['olaf_5fdb_5fmem_5faudio_5fid_1',['olaf_db_mem_audio_id',['../olaf__fp__ref__mem_8h.html#ac63617adfcff48859fd59c8f7c30b47a',1,'olaf_fp_ref_mem.h']]],
+  ['olaf_5fdb_5fmem_5faudio_5fpath_2',['olaf_db_mem_audio_path',['../olaf__fp__ref__mem_8h.html#ae5ee05e0bdf90fd0c6551fb2a1c65fab',1,'olaf_fp_ref_mem.h']]],
+  ['olaf_5fdb_5fmem_5ffp_5flength_3',['olaf_db_mem_fp_length',['../olaf__fp__ref__mem_8h.html#a02184939e71d0c18cef52046cec97deb',1,'olaf_fp_ref_mem.h']]],
+  ['olaf_5fdb_5fmem_5ffps_4',['olaf_db_mem_fps',['../olaf__fp__ref__mem_8h.html#ae61ccad003f3bb7d26ad1ce25e09d2cc',1,'olaf_fp_ref_mem.h']]]
 ];
