+++ src/olaf_fp_ref_mem.h
@@ -1,3428 +1,3283 @@
+/**@file olaf_fp_ref_mem.h 
+ * @brief The in-memory 'database' of fingerprints used to match queries against in the 'memory' configuration. 
+ */
 #include <stdint.h>
 
+/**The array with fingerprints.*/
 const uint64_t olaf_db_mem_fps[] = {
-5117709084,
-5117709564,
-5117711993,
-5117717794,
-5118171196,
-5118175036,
-5118237642,
-5118296938,
-5118305410,
-5118306302,
-5118306782,
-5118364039,
-5118371521,
-5118372961,
-5118492826,
-5118493306,
-5118559802,
-5118560282,
-5118561242,
-5118561722,
-5118563161,
-5118563641,
-5118564122,
-5118564602,
-5118565853,
-5118566521,
-5118567001,
-5118567481,
-5118569646,
-5118628577,
-5118629057,
-5118631937,
-5118632417,
-5118632897,
-5118698914,
-5118699394,
-5118700354,
-5118760569,
-5118761569,
-5118762049,
-5118886762,
-5118887242,
-5118888202,
-5118893002,
-5118960008,
-5119021495,
-5119021704,
-5119022455,
-5119025335,
-5119026295,
-5119157378,
-6174871541,
-6174872501,
-6174876341,
-6174939213,
-6175003302,
-6175006466,
-6175007426,
-6175008147,
-6175066918,
-6175067398,
-6175067907,
-6175069827,
-6175328187,
-6175329147,
-6175329627,
-6175336347,
-6175336827,
-6175393841,
-6175394801,
-6175395281,
-6175397681,
-6175398161,
-6175398641,
-6175402001,
-6175402481,
-6175457936,
-6175458416,
-6175460102,
-6175460582,
-6175461657,
-6175465977,
-6175467302,
-6175467782,
-6175526832,
-6175527312,
-6175530672,
-6175532757,
-6175533237,
-6175534197,
-6175661625,
-6175854581,
-6175859381,
-6175985411,
-6175987331,
-6175987789,
-6175992589,
-6176115462,
-6176115942,
-6176185788,
-7248423955,
-7248744180,
-7249734675,
-7249792756,
-7249798516,
-8340383004,
-14781581157,
-14781582597,
-14781583557,
-14781584517,
-14781587157,
-14781587637,
-14781588597,
-14782242517,
-14782242997,
-14782243957,
-16912231681,
-19059902414,
-23372569185,
-24429668254,
-24429668734,
-48050930702,
-49124221320,
-49124802621,
-49124803101,
-49125786141,
-49125789021,
-64156994026,
-64156994506,
-64156998826,
-64157001226,
-64157001706,
-64157002666,
-64157057162,
-64157061962,
-64157062442,
-64157065322,
-64157065802,
-64157066282,
-101755650217,
-101755650697,
-101755651177,
-101755652137,
-101755653097,
-101755653577,
-101755655977,
-101755656457,
-101755656937,
-117861781224,
-119991572100,
-119991635716,
-119991641476,
-128598155643,
-129656493639,
-130728863935,
-130728864415,
-131819445691,
-141483060021,
-142557850210,
-166179314236,
-181211507839,
-184433119960,
-194097187403,
-194097187883,
-194097192202,
-194097261578,
-237046268027,
-238119945756,
-238121004269,
-241341180602,
-250987875562,
-250987882762,
-279996270023,
-279996661793,
-279996989654,
-279996991094,
-281069750086,
-282127565525,
-282127566485,
-282127566965,
-283200398620,
-284274198058,
-284274458683,
-284274459163,
-284274460123,
-287513120485,
-293938732277,
-293938925746,
-297160016535,
-297160017015,
-299323359774,
-308986914526,
-308986915006,
-313266213481,
[…]
