+++ src/olaf_fp_ref_mem.h
@@ -5,99 +5,50 @@
 
 /**The array with fingerprints.*/
 const uint64_t olaf_db_mem_fps[] = {
-5117718415,
-5117845128,
-5118108125,
-5118430557,
 5118627614,
-5118635919,
-5118699918,
+5118896381,
 5118959223,
-5119025629,
-5119092483,
-6174936399,
-6174945177,
-6175003093,
-6175139205,
-6175397422,
-6175400910,
 6175534391,
-6175789525,
-6176058377,
-6176121806,
-6191580129,
-6191656458,
 6191974350,
-6192110175,
-6192112489,
 6192240529,
-7248683501,
 7248947520,
-7249136167,
-7249536264,
-7249593572,
+7249339757,
 7249728312,
-7265393913,
-7265592621,
+7249862442,
 7265717678,
-7265857216,
-7266379053,
-7266502213,
-7266569646,
-8323013561,
-8339529921,
-9395773625,
-9395962398,
+7265854745,
+7265977847,
+8323075052,
 9396160967,
-9396620875,
 9396694249,
-9396820362,
-9413197888,
-9413461316,
-9413534016,
+9413599680,
 9413602641,
-9414116676,
+9413986056,
 9414118631,
-9414187283,
-10469648211,
-10469836622,
-10469838592,
-10470097848,
+9414122014,
+9414187033,
 10470169511,
-10470303561,
 10470623904,
-10470690547,
 10470818429,
-10471082166,
-10471086572,
-10486556813,
+10487267906,
 10487343233,
 10487661973,
-11560954865,
 11561080517,
 12617460743,
 12617785068,
-12618116103,
-12633972077,
-12634033966,
-12634167067,
-12634499476,
+12617787545,
+12617787545,
 12634762816,
-13691068281,
-13691594322,
+12635407517,
+13691854506,
 13692115737,
+13708170286,
 13708301629,
-13708369272,
-13708373189,
-14765130851,
+13709149374,
 14765196275,
-14765270780,
-14765393119,
 14765665721,
-15839341653,
-15855322665,
 15855651270,
-16912746038,
+16912755845,
 16929136240,
 17986955954,
 19060235450,
@@ -105,3173 +56,2052 @@ const uint64_t olaf_db_mem_fps[] = {
 19060689069,
 19060829021,
 19076551222,
-19076813349,
+19076614280,
 19077269396,
-20133646759,
-20133779429,
-20133846023,
+19077665926,
 20133909341,
-20134039574,
-20134501373,
-20134561026,
-20134566919,
-20150688021,
+20133911530,
+20134168726,
+20134433747,
 21208499001,
-51271835615,
 52345907296,
-52346495864,
-55567254317,
+53436488976,
 56658103036,
 57714938717,
-57731912788,
-60952484888,
+58788746303,
 60953798044,
-62010562838,
 67395458923,
-72764489877,
-78116751427,
-83502042027,
-86722222906,
-86722416545,
-86723009326,
-92074611241,
-92090738984,
-93166046211,
-94222943680,
-95312679790,
-98517063657,
-99607055952,
-99607121411,
-101754741107,
-102812035101,
-113566616914,
-115697000046,
-115713909375,
+68469331840,
+86705833076,
+89927330509,
+91001000508,
+91017253937,
+91018370494,
+91018432962,
+94222945251,
+96370370406,
+98517059253,
+99607520600,
 117844872635,
 117861256113,
 120009198141,
-122156624092,
-123229573734,
-127525783260,
-131802859413,
-131803193653,
-136114546073,
+124287324610,
+131820692310,
+132877132606,
+133951455354,
 141467131847,
-145761632485,
 161884676560,
 167236864126,
-174769711894,
[…]

[…]
