### 2023-03-20 #5252 @JorenSix
To help with looking for reviewers I have checked the list of registered reviewers and have a couple more suggestions: Both expectopatronum and hadware have relevant audio DSP experience and C in their programming languages. FilipeChagasDev has experience with audio and embedded devices.
