### 2023-05-10 #5459 @editorialbot
[REVIEW]: Olaf: a lightweight, portable audio search system
**Submitting author:** @JorenSix (<a href="http://orcid.org/0000-0001-7671-1907">Joren Six</a>)
**Repository:** https://github.com/JorenSix/Olaf
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
