### 2023-05-10 JorenSix/Olaf#25 @liscio
Missing/limited instructions for running tests
I think it would be helpful if users/contributors had instructions for running the automated tests locally when possible. The README makes reference to the scripts in the `eval/` directory, but it's not clear how to use them, when contributors should execute certain tests, and so on.

Similarly, there appears to be a `test` target in the Makefile that claims to "build and test" olaf, but no test is actually executed. If this is intended to be part of the testing suite, it'd be a good idea to include that in the README as well.

Related to: https://github.com/openjournals/joss-reviews/issues/5459
