### 2023-05-12 JorenSix/Olaf#26 @JorenSix
Thanks for taking the time to try it out and test it.  Indeed the documentation for running the tests is lacking which gives room for confusion.

The @make test@ runs a couple of (outdated) sanity checks which were used during development. I will replace it with the functional tests (which include the unit tests in a way) and make sure it runs fine. It seems the error mainly has to do with not clearing the database properly before running tests.
