### 2023-05-12 f2715e4
Fixed unit tests, improved functional test with some more guardrails and checks, see #26
