+++ src/olaf_fp_ref_mem.h
@@ -5,2109 +5,2129 @@
 
 /**The array with fingerprints.*/
 const uint64_t olaf_db_mem_fps[] = {
-5118627614,
-5118896381,
-5118959223,
-6175534391,
-6191974350,
-6192240529,
-7248947520,
-7249339757,
-7249728312,
-7249862442,
-7265717678,
-7265854745,
-7265977847,
-8323075052,
-9396160967,
-9396694249,
-9413599680,
-9413602641,
-9413986056,
-9414118631,
-9414122014,
-9414187033,
-10470169511,
-10470623904,
-10470818429,
-10487267906,
-10487343233,
-10487661973,
-11561080517,
-12617460743,
-12617785068,
-12617787545,
-12617787545,
-12634762816,
-12635407517,
-13691854506,
-13692115737,
-13708170286,
-13708301629,
-13709149374,
-14765196275,
-14765665721,
-15855651270,
-16912755845,
-16929136240,
-17986955954,
-19060235450,
-19060430838,
-19060689069,
-19060829021,
-19076551222,
-19076614280,
-19077269396,
-19077665926,
-20133909341,
-20133911530,
-20134168726,
-20134433747,
-21208499001,
-52345907296,
-53436488976,
-56658103036,
-57714938717,
-58788746303,
-60953798044,
-67395458923,
-68469331840,
-86705833076,
-89927330509,
-91001000508,
-91017253937,
-91018370494,
-91018432962,
-94222945251,
-96370370406,
-98517059253,
-99607520600,
-117844872635,
-117861256113,
-120009198141,
-124287324610,
-131820692310,
-132877132606,
-133951455354,
-141467131847,
-161884676560,
-167236864126,
-170475392568,
-171549008562,
-182286560702,
-200539112491,
-205891579668,
-209129384665,
-215555246686,
-217719839165,
-219850344789,
-226292604017,
-230587902081,
-232752225220,
-235955749931,
-235973379205,
-242399195037,
-279995089185,
-279995089559,
-279995095063,
-279995099474,
-279995166030,
-279996211629,
-281070415159,
-282127176284,
-282143623517,
-282143754208,
-282143762752,
-282144025809,
-282144149302,
-282144543544,
-282144612873,
-283200391637,
-283217496533,
-283217768184,
-283217822249,
-284273280890,
-284273547658,
-284274202179,
-284274792990,
-284274854461,
-284275050951,
-284275057088,
-284275181434,
-284291108071,
-284291637100,
-285348207627,
-285348270663,
-285348473089,
-285348473128,
-285364651150,
-285364724860,
-285364852473,
-285364855716,
-285364920549,
-285364921919,
-285365052076,
-285365054486,
-285365119358,
-285365181186,
-285365250305,
-285365314338,
-287495294201,
-287495299320,
-287495496385,
-287496214237,
-287512595586,
-288569232413,
-288569239664,
-288569494099,
-288569495470,
-288569499724,
-288569499724,
-288569754741,
-288569764664,
-288570283054,
-288586542788,
-288586796590,
-289643957400,
-289659620504,
-289660021284,
-290716724377,
-290716978285,
-290734149217,
-290734279663,
-291807570637,
-292864527947,
-292864728514,
-292880984719,
-292881769564,
-292881899134,
-295011756704,
-295011946982,
-327241575234,
-329371878206,
-331536280416,
-333683561579,
-336904662039,
-339052400208,
-340125428101,
-361600391898,
-363731426073,
-366953371177,
-366969163547,
-366969297459,
-369100790940,
-377690007145,
-383076079876,
-398107749530,
-401312449982,
-401328778412,
-401329564253,
-402402846249,
[…]
