+++ src/pffft.h
@@ -185,7 +185,7 @@ extern "C" {
   /**
    * return 4 or 1 wether support SSE/Altivec instructions was enable when building pffft.c 
    */
-  int pffft_simd_size();
+  int pffft_simd_size(void);
 
 #ifdef __cplusplus
 }
