### 2023-05-12 JorenSix/Olaf#26 @JorenSix
The problem had indeed to do with a non-empty database when starting the functional tests (e.g. when a previous test is aborted). Now there is a warning if a non-empty database is detected.

The database is not thrashed automatically since it can contain a large amounts of user data.
