### 2023-05-12 JorenSix/Olaf#25 @JorenSix
The readme has been extended with instructions to run the unit tests and the functional tests.
The readme is already a bit long to my tastes. Perhaps it needs some more attention to structure it better (the order of headers is a bit random now...).
