### 2023-05-15 #5459 @JorenSix
Just an update from here, I have fixed and closed the reported issues. They were mainly about fixing instructions and offering some proper error messages when things (tests) go wrong.

I have also updated the readme. The text grew over time and is now more properly structured. This is needed since there are instructions for running Olaf on traditional computers, embedded devices and in the browser.  I have also added an index which helps to navigate the (perhaps long) document.
