### 2023-06-08 3c668b6
Improved command line interface with help messages and fixed 'olaf mic' command. Readme reflects the changes
