+++ olaf.rb
@@ -28,7 +28,6 @@ EXECUTABLE_LOCATION = "/usr/local/bin/olaf_c"
 CHECK_INCOMING_AUDIO = true
 SKIP_DUPLICATES = true
 MONITOR_LENGTH_IN_SECONDS = 7
-AVAILABLE_THREADS = 7 #change to e.g. your number of CPU cores -1
 TARGET_SAMPLE_RATE = 16000
 
 ALLOWED_AUDIO_FILE_EXTENSIONS = "**/*.{m4a,wav,mp4,wv,ape,ogg,mp3,raw,flac,wma,M4A,WAV,MP4,WV,APE,OGG,MP3,FLAC,WMA}"
@@ -45,397 +44,398 @@ MIC_INPUT = "ffmpeg -hide_banner -loglevel panic  -f avfoundation -i 'none:defau
 # a text file is read and each line is interpreted as a path to a file
 # for a folder each audio file within that folder (and subfolders) is added to the list
 def audio_file_list(arg,files_to_process)
-	arg = File.expand_path(arg)
-	if File.directory?(arg)
-		audio_files_in_dir = Dir.glob(File.join(arg,ALLOWED_AUDIO_FILE_EXTENSIONS))
-		audio_files_in_dir.each do |audio_filename|
-			files_to_process << audio_filename
-		end
-	elsif File.extname(arg).eql? ".txt"
-		audio_files_in_txt = File.read(arg).split("\n")
-		audio_files_in_txt.each do |audio_filename|
-			audio_filename = File.expand_path(audio_filename)
-			if File.exist?(audio_filename)
-				files_to_process << audio_filename
-			else
-				STDERR.puts "Could not find: #{audio_filename}"
-			end
-		end
-	elsif File.exist? arg
-		files_to_process << arg
-	else
-		STDERR.puts "Could not find: #{arg}"
-	end
-	files_to_process
+  arg = File.expand_path(arg)
+  if File.directory?(arg)
+    audio_files_in_dir = Dir.glob(File.join(arg,ALLOWED_AUDIO_FILE_EXTENSIONS))
+    audio_files_in_dir.each do |audio_filename|
+      files_to_process << audio_filename
+    end
+  elsif File.extname(arg).eql? ".txt"
+    audio_files_in_txt = File.read(arg).split("\n")
+    audio_files_in_txt.each do |audio_filename|
+      audio_filename = File.expand_path(audio_filename)
+      if File.exist?(audio_filename)
+        files_to_process << audio_filename
+      else
+        STDERR.puts "Could not find: #{audio_filename}"
+      end
+    end
+  elsif File.exist? arg
+    files_to_process << arg
+  else
+    STDERR.puts "Could not find: #{arg}"
+  end
+  files_to_process
 end
 
 #Folder size in MB
 def folder_size(folder)
-	folder = folder
-	total_size = 0
-	Find.find(folder) do |path|
-	  if File.file?(path)
-	    total_size += File.size(path)
-	  end
-	end
-	total_size / (1024.0 * 1024.0)
+  folder = folder
+  total_size = 0
+  Find.find(folder) do |path|
+    if File.file?(path)
+      total_size += File.size(path)
+    end
+  end
+  total_size / (1024.0 * 1024.0)
 end
 
 
 def has(audio_file)
-	#return false if no db exits
-	return false if (Dir.glob(File.join(DB_FOLDER,"*")).size == 0)
-	
-	result = `#{EXECUTABLE_LOCATION} has '#{audio_file}'`
-	unless(result.empty?)
-		result_line = result.split("\n")[1]
-		result_line.split(";").size > 1
-	else
-		false
-	end
+  #return false if no db exits
[…]
