### 2023-06-08 c3c73ca
Fix checking valid result lines
