### 2023-06-09 #5459 @JorenSix
@ebezzam thanks for taking the time for the review!

* Indeed there is some overlap with the Panako JOSS paper, as I disclosed at the submission stage. The two projects have similarities but each have their own niche (time/frequency robustness in Panako, embedded use in Olaf). The similarities explain the overlap in the more general intro and similar references.  The referenced  Olaf ISMIR LBD abstract is not peer-reviewed and serves a different purpose than this paper: the LBD abstract is more technical and aimed at domain experts, while this paper introduces Olaf to a more general audience and - crucially - is peer reviewed ;).
* Measuring memory use is indeed a good idea. I have already started using valgrind to get insights and identified a memory leak in the use of the hash table! The memory use depends on the size of the database but is below 3MB in the case of using the test dataset. More on this later..
*  A performance comparison seems like a good idea. I can run the same benchmark for both system and report them.
* Chromaprint serves another case: it creates a fingerprint of a full track and only matches full duplicates of tracks. It is not designed for granular queries. I can add this to text to clarify this.
* The minor comment is noted. I am not sure about the best way to reference software in latex but I will look into it.
