### 2023-06-09 #5459 @bmcfee
Thanks @ebezzam !

> * I don't know how this fits with [JOSS ethics](https://joss.theoj.org/about#code_of_conduct) on self-plagiarism. Maybe @bmcfee can clarify?

I think this is fine and in accordance with the ethics guidelines.  As @JorenSix notes, ISMIR Late-breaking demos (LBDs) are not peer-reviewed and are not published in the proceedings.  It's quite common for preliminary work to first appear as an LBD and later be published properly (at ISMIR or elsewhere).

> The minor comment is noted. I am not sure about the best way to reference software in latex but I will look into it.

If the software being cited does not have a DOI, the JOSS submission guidelines give an example of how to do this: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography - excerpt:

```
@misc{fidgit,
  author = {A. M. Smith and K. Thaney and M. Hahnel},
  title = {Fidgit: An ungodly union of GitHub and Figshare},
  year = {2020},
  publisher = {GitHub},
  journal = {GitHub repository},
  url = {https://github.com/arfon/fidgit}
}
```
