### 2023-06-12 9aaa113
Memory leak fixed in the hash table use: called free instead of hash_table_free.
