### 2023-06-12 af354bf
Reduced memory usage by correctly using the hash table and not having an additional ever growing list of match results. This also simplifies the implementation.
