+++ src/olaf_fp_matcher.c
@@ -54,15 +54,6 @@ inline int min ( int a, int b ) { return a < b ? a : b; }
 
 struct Olaf_FP_Matcher{
 
-	//List with all results
-	struct match_result * m_results;
-
-	//the current size of the results found
-	size_t m_results_size;
-
-	//the current index in the array
-	size_t m_results_index;
-
 	//A hash table to quickly check whether a match_id and time diff
 	//combination has already been encountered
 
@@ -107,143 +98,54 @@ int uint64_t_equal(void *vlocation1, void *vlocation2){
 	return *location1 == *location2;
 }
 
+
+//void olaf_hash_table_key_free_func(void *key){
+	//NOOP, key is freed by value free func
+	//
+	//fprintf(stderr,"%llu free key\n", *((uint64_t *)key));
+	//free(key);
+//}
+
+void olaf_hash_table_value_free_func(void *value){
+	free(value);
+}
+
 //Creates a new matcher 
 Olaf_FP_Matcher * olaf_fp_matcher_new(Olaf_Config * config,Olaf_DB* db ){
 	Olaf_FP_Matcher *fp_matcher = (Olaf_FP_Matcher *) malloc(sizeof(Olaf_FP_Matcher));
-
+	
+	//The database results are integers which combine a time info and match id
+	fp_matcher->db_results = (uint64_t *) calloc(config->maxDBCollisions , sizeof(uint64_t));
+	fp_matcher->result_hash_table = hash_table_new(uint64_t_hash,uint64_t_equal);
+	fp_matcher->last_print_at = 0;
+	fp_matcher->config = config;
 	fp_matcher->db = db;
 
-	//The maximum number of results returned by the database
-	//This is the number of times a 'hash collision' is allowed in the 
-	//database
-	fp_matcher->m_results_size =  config->maxDBCollisions;
-
-	//The database results are integers which combine a time info and match id
-	fp_matcher->db_results = (uint64_t *) calloc(fp_matcher->m_results_size , sizeof(uint64_t));
-	fp_matcher->config = config;
-	fp_matcher->result_hash_table = hash_table_new(uint64_t_hash,uint64_t_equal);
-	fp_matcher->m_results = (struct match_result *) calloc(fp_matcher->m_results_size , sizeof(struct match_result));
-	fp_matcher->m_results_index = 0;
-	fp_matcher->last_print_at = 0;
+	hash_table_register_free_functions(fp_matcher->result_hash_table,NULL, olaf_hash_table_value_free_func);
 
 	return fp_matcher;
 }
 
-//for use with qsort: a comparator that sorts result structs by age
-//newest first!
-int olaf_fp_sort_results_by_age(const void * a, const void * b) {
-	struct match_result * aResult = (struct match_result*)a;
-	struct match_result * bResult = (struct match_result*)b;
-
-	//first sort by match count from newest to oldest
-	int diff = (bResult->queryFingerprintT1 - aResult->queryFingerprintT1);
-
-	return diff;
-}
-
-//This removes old matches for streaming purposes.
-//It does a couple of things:
-// 1 sort all matches from new to old
-// 2 reset all matches which are too old
-void olaf_fp_matcher_mark_old_matches(Olaf_FP_Matcher * fp_matcher, int current_query_time ){
-	
-	//sort matches by age
-	if(fp_matcher->m_results_index > 0)
-		qsort(fp_matcher->m_results, fp_matcher->m_results_size, sizeof(struct match_result), olaf_fp_sort_results_by_age);
+// This iterates the hash table with matches and 
+// removes old matches for streaming purposes reset all matches which are too old
[…]
