### 2023-06-15 e04daeb
Fix for casting warning
