+++ paper.bib
@@ -367,3 +367,67 @@
   url = {https://doi.org/10.1007/978-3-031-18444-4_16},
   year = {2023}}
 
+
+
+@misc{calgorithms,
+  author = {Simon Howard},
+  title = {C Algorithms: a collection of common Computer Science algorithms},
+  year = {2020},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/fragglet/c-algorithms}
+}
+
+@misc{audfprint,
+  author = {Dan Ellis},
+  title = {audfprint: Landmark-based audio fingerprinting},
+  year = {2019},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/dpwe/audfprint}
+}
+
+@misc{pffft,
+  author = {Julien Pommier},
+  title = {pffft:  a pretty fast FFT},
+  year = {2022},
+  publisher = {BitBucket},
+  journal = {BitBucket repository},
+  url = {https://bitbucket.org/jpommier/pffft/src/master/}
+}
+
+@misc{zig,
+  author = {Andrew Kelley and contributors},
+  title = {zig: a general-purpose programming language and toolchain for maintaining robust, optimal, and reusable software.},
+  year = {2023},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/ziglang/zig}
+}
+
+@misc{lmdb,
+  author = {Howard Chu},
+  title = {lmdb: a general-purpose programming language and toolchain for maintaining robust, optimal, and reusable software.},
+  year = {2022},
+  publisher = {OpenLDAP},
+  journal = {OpenLDAP repository},
+  url = {https://git.openldap.org/openldap/openldap/tree/mdb.master}
+}
+
+@misc{doxygen,
+  author = {Dimitri van Heesch},
+  title = {Doxygen: the de facto standard tool for generating documentation from annotated C++ sources.},
+  year = {2023},
+  publisher = {Doxygen},
+  journal = {Doxygen website},
+  url = {https://www.doxygen.nl/}
+}
+
+@misc{chromaprint,
+  author = {Lukáš Lalinský and contributors},
+  title = {Chromaprint: an audio fingerprint library developed for the AcoustID project.},
+  year = {2023},
+  publisher = {Github},
+  journal = {GitHub repository},
+  url = {https://github.com/acoustid/chromaprint}
+}
