### 2023-06-15 0a77207
Olaf vs Panako script + memory use script fixes
