### 2023-06-15 JorenSix/Olaf#28 @ebezzam
Small typos in README
Related to https://github.com/openjournals/joss-reviews/issues/5459

https://github.com/JorenSix/Olaf/blob/0a772074729bd9d5e1d27d934a75b37417d6d9af/README.textile?plain=1#L248

- "whether"
- "...works. The..."

https://github.com/JorenSix/Olaf/blob/0a772074729bd9d5e1d27d934a75b37417d6d9af/README.textile?plain=1#L89

- "also"

https://github.com/JorenSix/Olaf/blob/0a772074729bd9d5e1d27d934a75b37417d6d9af/README.textile?plain=1#L261

- fix code formatting

https://github.com/JorenSix/Olaf/blob/0a772074729bd9d5e1d27d934a75b37417d6d9af/README.textile?plain=1#L159

- add code formatting here and fix formatting in the list below

https://github.com/JorenSix/Olaf/blob/0a772074729bd9d5e1d27d934a75b37417d6d9af/README.textile?plain=1#L188

- "coming"
