### 2023-06-15 JorenSix/Olaf#30 @ebezzam
Units tests "error"?
Related to https://github.com/openjournals/joss-reviews/issues/5459

I think it's normal, but when running the unit tests on Linux (Ubuntu 22.04 LTS) and macOS (Catalina):
```
git clone https://github.com/JorenSix/Olaf
cd Olaf
make && make install
ruby eval/olaf_functional_tests.rb
```
 I get the following "error" at the very beginning:
```
Database Error: No such file or directory
```
But I guess this is normal as there is a "PASS" just after to make sure the Database is empty?

Otherwise all tests pass 👍
