### 2023-06-15 JorenSix/Olaf#31 @ebezzam
Clearer install
Related to https://github.com/openjournals/joss-reviews/issues/5459

https://github.com/JorenSix/Olaf/blob/0a772074729bd9d5e1d27d934a75b37417d6d9af/README.textile?plain=1#L64

You do it later on for the tests, but perhaps you could also give the full installation steps even if it's "obvious":
```
sudo apt-get install ffmpeg ruby
git clone https://github.com/JorenSix/Olaf
cd Olaf
make && make install
```

On a similar note, is it possible to make `Olaf` ["apt-get install"-able](https://askubuntu.com/questions/336130/create-apt-get-install-for-own-software)? And such that dependencies are also installed. That would be nice for convenience.
