### 2023-06-15 JorenSix/Olaf#32 @ebezzam
Clear database
Related to https://github.com/openjournals/joss-reviews/issues/5459

https://github.com/JorenSix/Olaf/blob/0a772074729bd9d5e1d27d934a75b37417d6d9af/README.textile?plain=1#L201

Could it also be worth mention the `olaf clear` command in this section?
