### 2023-06-15 JorenSix/Olaf#33 @ebezzam
Formatting
Related to https://github.com/openjournals/joss-reviews/issues/5459

https://github.com/JorenSix/Olaf/blob/0a772074729bd9d5e1d27d934a75b37417d6d9af/README.textile?plain=1#L39

Perhaps you can write out the three reasons as a list so that it stands out?
