### 2023-06-15 #5459 @ebezzam
Thanks @JorenSix and @bmcfee for the quick responses and clarifications. I've opened a few issues on minor things / quick fixes.

On the unchecked points of my review:
- [ ] **Performance**: (1) low memory and (2) comparison with Panako would be great.
- [ ] **Installation instruction**: comments in https://github.com/JorenSix/Olaf/issues/31. Moreover, a package manager option would be great.
- [ ] **Example usage**: I think a dedicated "Examples" section could be nice in the README, where you provide a dummy folder of audio files, as you do for the tests.
- [ ] **State of field:** While Chromaprint serves another case, as you mentioned in your response, I think it's interesting to make that distinction. Other relevant software that a quick Google search reveals are [dejavu](https://github.com/worldveil/dejavu) and [SpectroMap](https://github.com/Aaron-AALG/SpectroMap).
- [ ] **References**: having the footnotes as references.

Feature request, which I think could really enhance usability and usage, is a Python wrapper. Like [this](https://github.com/rabitt/pysox) for SoX. I'll let @bmcfee weigh in if he thinks this is necessary for acceptance.

Hope that helps!
