### 2023-06-15 #5459 @bmcfee
Thanks @ebezzam !

I don't think a python wrapper is necessary here, though it would of course be a nice feature at some point.
