+++ eval/olaf_evaluation.rb
@@ -5,7 +5,7 @@ require 'tmpdir'
 ALLOWED_AUDIO_FILE_EXTENSIONS = "**/*.{m4a,wav,mp4,wv,ape,ogg,mp3,flac,wma,M4A,WAV,MP4,WV,APE,OGG,MP3,FLAC,WMA}"
 AUDIO_FILES_TO_CHECK_FOR_TRUE_POSITIVES = 100
 AUDIO_FILES_TO_CHECK_FOR_FALSE_POSITIVES = 100
-AUDIO_FILES_OTHERS = 400
+AUDIO_FILES_OTHERS = 100
 QUERY_LENGTHS = [10]
 RANDOM_SEED = 1
 
@@ -228,8 +228,7 @@ def print_olaf_query_result(query_file,modification,parameter,ref_file_start,mat
 
   #match count (#), q start (s) , q stop (s), ref path, ref ID, ref start (s), ref stop (s)
   #store only first match
-
-  score,query_start,query_stop,ref_path,ref_id,ref_start,ref_stop = line.split(",")
+  indx,count,query_file,score ,query_start,query_stop,ref_path,ref_id,ref_start,ref_stop = line.split(",").map(&:strip)
   m = Match.new
 
   m.score = score.to_i
@@ -240,6 +239,9 @@ def print_olaf_query_result(query_file,modification,parameter,ref_file_start,mat
   m.time_diff = ref_start.to_f - query_start.to_f
   m.match_expected = ("tp"==tp_or_tn_expected)
   m.query_file = query_file
+
+  puts "line: #{line} "
+  puts "match: #{m.file_names_match?} #{m.query_file} #{m.result_file_name}"
 
   matches << m
 
