### 2023-06-16 e67f471
Extended install instructions, see #31
