### 2023-06-16 429d244
Hide database error message: it should be empty the first time. See #30
