### 2023-06-16 JorenSix/Olaf#31 @JorenSix
I have extended the install instructions and made them more copy paste friendly. I landed on the following:

<pre>
#sudo apt-get install ffmpeg ruby
git clone https://github.com/JorenSix/Olaf
cd Olaf
make
sudo make install
</pre>

Packaging (apt, yum, eopkg, snap, brew, ...) is certainly a nice to have. Olaf is relatively simple and should be easy to package but I have not done this before and should read into it. However, packaging does add an additional maintenance burden and with Olaf being simple to @make && make install@ it is currently not high on my priority list. If there are packaging volunteers.. do not hesitate to let me know.
