### 2023-06-16 JorenSix/Olaf#32 @JorenSix
Thanks for the remark.

Indeed it was an oversight and now it is mentioned.
