### 2023-06-16 218acc6
Typos and formatting fixes, see #33 and #28
