### 2023-06-16 JorenSix/Olaf#28 @JorenSix
Thanks for the pointers.

They have been fixed together with some more typo's after running a spell-checker on the text!
