+++ paper.bib
@@ -431,3 +431,32 @@
   journal = {GitHub repository},
   url = {https://github.com/acoustid/chromaprint}
 }
+
+@misc{dejavu,
+  author = {Will Drevo},
+  title = {Dejavu: Audio fingerprinting and recognition algorithm implemented in Python.},
+  year = {2020},
+  publisher = {Github},
+  journal = {GitHub repository},
+  url = {https://github.com/worldveil/dejavu}
+}
+
+@InProceedings{spectromap,
+author="L{\'o}pez-Garc{\'i}a, Aar{\'o}n
+and Mart{\'i}nez-Rodr{\'i}guez, Brian
+and Liern, Vicente",
+editor="Montiel, Mariana
+and Agust{\'i}n-Aquino, Octavio A.
+and G{\'o}mez, Francisco
+and Kastine, Jeremy
+and Lluis-Puebla, Emilio
+and Milam, Brent",
+title="A Proposal to Compare the Similarity Between Musical Products. One More Step for Automated Plagiarism Detection?",
+booktitle="Mathematics and Computation in Music",
+year="2022",
+publisher="Springer International Publishing",
+address="Cham",
+pages="192--204",
+doi          = {{10.1007/978-3-031-07015-0_16}},
+}
+
