### 2023-06-16 #5459 @JorenSix
Thanks for the feedback! I have been busy and Olaf has improved quite a bit!  The changes:

* I have measured memory use and reported memory use [on a separate page](https://github.com/JorenSix/Olaf/blob/master/eval/README.textile). The readme is already long so I think the separate page makes sense. I have fixed a memory leak in the hashtable use which came to my attention during the memory use measurement. It also simplified the matcher [implementation](https://github.com/JorenSix/Olaf/blob/master/src/olaf_fp_matcher.c)
* I have compared the performance of [Panako and Olaf directly](https://github.com/JorenSix/Olaf/blob/master/eval/README.textile). Surprisingly the Java implementation (Panako) is slightly faster on my machine but this makes sense considering the fact that Panako is multi-threaded and has a much higher memory use. Also Java is not that slow in pure number crunching.
* The same page (eval readme) also contains info on the evaluation script which has been updated as well.
* I have improved the [readme](https://github.com/JorenSix/Olaf) fixing several typos and included a copy-pastable example use. Also the installation instructions are more clear now and include a complete set of instructions.
* The paper has been improved by additional references to audio fingerprinting systems and a note on the memory use. The reference to software are done properly now.
* The functional tests do not show a confusing error any more.

I agree that a python wrapper and packaging (for apt, yum, pacman, ...) are two nice-to-haves which would increase adoption of Olaf but I see this as separate from the core Olaf software (as is the case for the pysox).

Thanks again for taking the time to review this paper/software and please let me know if there are additional points to tackle!
