### 2023-06-20 47f7985
download over the air query set
