### 2023-06-20 45acfa9
Index out of bounds check
