### 2023-06-20 0173ff1
shorter mem db
