+++ src/olaf_fp_ref_mem.h
@@ -5,2129 +5,491 @@
 
 /**The array with fingerprints.*/
 const uint64_t olaf_db_mem_fps[] = {
-5118305822,
-5118306302,
-5118306782,
-5118306945,
-5118307262,
-5118371521,
-5118372001,
-5118372961,
-5118493786,
-5118559802,
-5118560282,
-5118561242,
-5118561722,
-5118562013,
-5118562201,
-5118562681,
-5118563161,
-5118563641,
-5118564122,
-5118564602,
-5118566521,
-5118567001,
-5118567481,
-5118628097,
-5118628577,
-5118629057,
-5118631457,
-5118631937,
-5118632417,
-5118632897,
-5118698914,
-5118699394,
-5118699874,
-5118700354,
-5118757729,
-5118758209,
-5118760569,
-5118760609,
-5118761089,
-5118761569,
-5118886282,
-5118886762,
-5118888202,
-5118888682,
-5118893002,
-5118956213,
-5119021704,
-6174938771,
-6175002626,
-6175003106,
-6175003302,
-6175003586,
-6175004066,
-6175006466,
-6175006946,
-6175007906,
-6175328667,
-6175329147,
-6175332027,
-6175332507,
-6175333467,
-6175335867,
-6175393361,
-6175394321,
-6175395281,
-6175397681,
-6175398161,
-6175401521,
-6175402001,
-6175402481,
-6175402961,
-6175457456,
-6175457936,
-6175458416,
-6175461657,
-6175462256,
-6175462617,
-6175463942,
-6175465977,
-6175466457,
-6175467782,
-6175526832,
-6175528272,
-6175530672,
-6175531152,
-6175531632,
-6175532112,
-6175532757,
-6175533237,
-6175534197,
-6175534277,
-6175599252,
-6175661625,
-6175662105,
-6175987789,
-6175988749,
-6175991629,
-6175992589,
-6176114982,
-6176115942,
-7248745170,
-14781587157,
-14781588117,
-19059902414,
-45920223430,
-45920223910,
-48050929742,
-48050930222,
-49124802621,
-49124803101,
-49124803581,
-49125785661,
-49125786141,
-49125792381,
-49125792861,
-64156994026,
-64156994506,
-64156994986,
-64156997386,
-64156997866,
-64156998346,
-64156998826,
-64157001226,
-64157001706,
-64157002186,
-64157002666,
-64157058122,
-64157059082,
-78132478510,
-117861781224,
-119991572100,
-119991635716,
-119991639556,
-119991641476,
-129656493639,
-166179314236,
-194097195082,
-194097261578,
-203760801546,
-204817702261,
-237046268027,
-238119945756,
-262816012508,
-279996270023,
-279996661793,
-279996991574,
-281069750086,
-282127566965,
-283200456714,
-283201308554,
-284274198058,
-284274458683,
-284274459163,
-284274459643,
-287513120485,
-293938732277,
-297160016055,
-297160016535,
-297160017015,
-297160017495,
-337978983513,
-348716468818,
-348716469298,
-348716473138,
-412050391433,
-412050391913,
-413140784175,
-413140784175,
-413140848751,
-418508645673,
-468975036011,
-468975619814,
-479712776097,
-479712781497,
-518349851339,
-521587858122,
-534472691579,
-560225459735,
-560225846046,
-560225846526,
-560225847006,
-560226042174,
-562373267173,
-564521082371,
-565594817770,
-565594818250,
-586012361068,
-586012361548,
-586012362988,
-586012363468,
-586012366828,
-586012367788,
-586012368268,
-586012370668,
-586012371148,
-590290555003,
-590290555963,
-612855913042,
-612855916882,
-612856890787,
-612856892707,
-612856895587,
-612856899427,
-612856900387,
[…]
