### 2023-06-21 JorenSix/Olaf#35 @ebezzam
Missing square brackets around references.
Related to: https://github.com/openjournals/joss-reviews/issues/5459

https://github.com/JorenSix/Olaf/blob/0173ff18411294ea0577569e6cbd52dbad8fd85d/paper.md?plain=1#L34

https://github.com/JorenSix/Olaf/blob/0173ff18411294ea0577569e6cbd52dbad8fd85d/paper.md?plain=1#L40
