### 2023-06-21 JorenSix/Olaf#36 @ebezzam
On performance / low memory
Related to: https://github.com/openjournals/joss-reviews/issues/5459

https://github.com/JorenSix/Olaf/blob/0173ff18411294ea0577569e6cbd52dbad8fd85d/paper.md?plain=1#L38

- Typo "...of Olaf allow it..."
- I would bring the footnote content to the main text, and add a hyperlink to the script you mention.
