### 2023-06-21 JorenSix/Olaf#37 @ebezzam
Mention comparison to Panako
Related to: https://github.com/openjournals/joss-reviews/issues/5459

https://github.com/JorenSix/Olaf/blob/0173ff18411294ea0577569e6cbd52dbad8fd85d/paper.md?plain=1#L40

- could be interesting to mention and link to the Panako comparison [you did](https://github.com/JorenSix/Olaf/blob/master/eval/README.textile#benchmark--how-fast-does-olaf-run-vs-panako)
