### 2023-06-21 JorenSix/Olaf#38 @ebezzam
Typos in eval page
Related to: https://github.com/openjournals/joss-reviews/issues/5459

Could you run a spell-checker on this page? There a few typos. For example:

https://github.com/JorenSix/Olaf/blob/0173ff18411294ea0577569e6cbd52dbad8fd85d/eval/README.textile?plain=1#L3

- "whether"
- missing space after first sentence

https://github.com/JorenSix/Olaf/blob/0173ff18411294ea0577569e6cbd52dbad8fd85d/eval/README.textile?plain=1#L51

- incomplete sentence

https://github.com/JorenSix/Olaf/blob/0173ff18411294ea0577569e6cbd52dbad8fd85d/eval/README.textile?plain=1#L84

- "whether"
