+++ src/pffft.h
@@ -183,7 +183,7 @@ extern "C" {
   void pffft_aligned_free(void *);
 
   /**
-   * return 4 or 1 wether support SSE/Altivec instructions was enable when building pffft.c 
+   * return 4 or 1 whether support SSE/Altivec instructions was enable when building pffft.c 
    */
   int pffft_simd_size(void);
 
