### 2023-06-21 fc8243c
Mem config fix
