+++ src/olaf_fp_ref_mem.h
@@ -5,491 +5,2129 @@
 
 /**The array with fingerprints.*/
 const uint64_t olaf_db_mem_fps[] = {
-5118625767,
-5118822230,
-5119017013,
-6192103873,
-7248938905,
-7249724351,
-7265911665,
-10469901084,
-10487269082,
-12617844465,
-12617844465,
-12634556396,
-17987012874,
-20134300060,
-52345898681,
-52346487248,
-93165258612,
-96370296255,
-99607511984,
-102812222656,
-131820618159,
-132877123991,
-174769703278,
-279995090859,
-282143491953,
-282143754137,
-283218346747,
-285348464513,
-285364716244,
-285364913304,
-285365043460,
-285365108785,
-285365502251,
-287495423818,
-287495553305,
-288569100146,
-288569296584,
-291807496486,
-292881041639,
-292881695413,
-292881893757,
-295011682553,
-295011746588,
-327241501083,
-342273164004,
-368025994399,
-401329555637,
-402402837634,
-468958315409,
-497965925021,
-505465800409,
-554873128790,
-554873193948,
-554873323773,
-554873455759,
-554873455759,
-554873979909,
-554874504319,
-554874505616,
-554874963194,
-557022118847,
-560225058865,
-560226567960,
-560226764058,
-562373527532,
-562373788388,
-562374049863,
-562390302791,
-562390762458,
-568816566710,
-613929519274,
-636478424821,
-641830946522,
-641847329855,
-643995467819,
-681575383113,
-707345514537,
-717009715468,
-773917312968,
-776048542386,
-786802280425,
-829751101415,
-829751951413,
-829752280883,
-829752412135,
-829752606973,
-829752607887,
-829752607887,
-829752608598,
-830825628097,
-831899698033,
-832972260138,
-835120661276,
-835120924881,
-836176708258,
-836194468514,
-837250384858,
-837250973767,
-837250975562,
-837251498970,
-837268078873,
-838325175467,
-838341691250,
-838341887688,
-841546401938,
-842620928693,
-843693689163,
-843694737723,
-844768280874,
-845841236022,
-845841367761,
-845841368216,
-845841696724,
-945716069807,
-946790402027,
-954289292004,
-954289423628,
-972542903609,
-1000460191032,
-1015509484190,
-1019787478772,
-1019805107147,
-1109997715505,
-1109998372122,
-1109998568852,
-1112146249457,
-1112146249457,
-1115350107059,
-1118571136273,
-1118572512508,
-1119644420175,
-1119644551207,
-1119644616079,
-1119645141404,
-1119645335796,
-1120718226129,
-1120718227412,
-1120719210246,
-1120719537882,
-1166890567660,
-1174407219133,
-1192677672095,
-1215225725596,
-1227036885891,
-1229167461588,
-1231315010710,
-1231315010710,
-1245273720870,
-1301124483052,
-1327951251189,
-1379507242091,
-1380580459046,
-1382729974523,
-1391301624949,
-1391302673525,
-1392375957885,
-1393449173430,
-1393450484150,
-1394523112220,
-1454653047667,
-1464333370542,
-1469702867305,
-1471833310292,
-1560971118091,
-1578149938427,
-1593182258889,
-1598552082376,
-1654384624890,
-1654385281424,
+5118305822,
+5118306302,
+5118306782,
+5118306945,
+5118307262,
+5118371521,
+5118372001,
+5118372961,
+5118493786,
+5118559802,
+5118560282,
+5118561242,
+5118561722,
+5118562013,
+5118562201,
+5118562681,
+5118563161,
+5118563641,
+5118564122,
+5118564602,
+5118566521,
[…]
