### 2023-06-21 JorenSix/Olaf#35 @JorenSix
There is a deliberate formatting difference when using square brackets or not. In some cases the references are meant to fit in the text.

For example "a paper by Fenet et al (2020)" is without brackets while "there are a number of papers on blablabla [Fenet 2020; Wang 2003]" is with brackets.
