### 2023-06-21 JorenSix/Olaf#38 @JorenSix
I have ran it through a spell checker and fixed some copy paste errors.
