### 2023-06-22 6257852
Implemented the van herk max filter to improve naive max filtering, resulting in about a 4x speedup in fingerprint extraction
