+++ src/olaf_max_filter_perceptual_van_herk.c
@@ -1,0 +1,121 @@
+// Olaf: Overly Lightweight Acoustic Fingerprinting
+// Copyright (C) 2019-2023  Joren Six
+
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+#include <assert.h>
+
+#include "olaf_max_filter.h"
+
+//these are computed to allow +-1000 cents (almost an octave)
+//the first zeros mean that the lower bins are ignored up until about 110Hz
+//This assumes also 16kHz sample rate
+const size_t olaf_max_filter_perceptual_min_idx[] = {0,0,0,0,0,0,0,0,0,9,9,9,9,9,9,9,9,10,10,11,12,12,12,13,14,14,14,15,15,16,16,17,17,18,19,19,19,21,21,22,22,23,23,25,25,25,26,26,26,27,27,27,29,29,29,31,31,31,33,33,33,35,35,35,35,37,37,37,37,39,39,39,39,41,41,41,41,43,43,43,43,43,47,47,47,47,47,51,51,51,51,51,53,53,53,53,53,55,55,55,55,55,55,59,59,59,59,59,59,63,63,63,63,63,63,63,67,67,67,67,67,67,67,71,71,71,71,71,71,71,75,75,75,75,75,75,75,75,79,79,79,79,79,79,79,79,83,83,83,83,83,83,83,83,83,87,87,87,87,87,87,87,87,87,95,95,95,95,95,95,95,95,95,95,99,99,99,99,99,99,99,99,99,99,103,103,103,103,103,103,103,103,103,103,103,111,111,111,111,111,111,111,111,111,111,111,111,119,119,119,119,119,119,119,119,119,119,119,119,127,127,127,127,127,127,127,127,127,127,127,127,127,135,135,135,135,135,135,135,135,135,135,135,135,135,135,143,143,143,143,143,143,143,143,143,143,143,143,143,143,151,151,151,151,151,151,151,151,151,151,151,151,151,151,151,151,159,159,159,159,159,159,159,159,159,159,159,159,159,159,159,159,167,167,167,167,167,167,167,167,167,167,167,167,167,167,167,167,167,167,175,175,175,175,175,175,175,175,175,175,175,175,175,175,175,175,175,175,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,199,199,199,199,199,199,199,199,199,199,199,199,199,199,199,199,199,199,199,199,199,207,207,207,207,207,207,207,207,207,207,207,207,207,207,207,207,207,207,207,207,207,207,223,223,223,223,223,223,223,223,223,223,223,223,223,223,223,223,223,223,223,223,223,223,223,239,239,239,239,239,239,239,239,239,239,239,239,239,239,239,239,239,239,239,239,239,239,239,239,239,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,271,271,271,271,271,271,271,271,271,271,271,271,271,271,271,271,271,271,271,271,271,271,271,271,271,271,271,287,287,287,287,287,287,287,287,287,287,287,287,287,287,287,287,287,287,287,287,287};
[…]
