### 2023-06-22 #5459 @JorenSix
Just an FYI, I have improved the speed of Olaf significantly with an algorithmic optimisation in the max-filter step. After a profiling session It became clear that this was the main speed bottleneck. Now it implements a (partial) Van Herk max filter as described in Van Herk, M. (1992). _A fast algorithm for local minimum and maximum filters on rectangular and octagonal kernels_. Pattern Recognition Letters, 13(7), 517-521.

The Panako vs Olaf results have changed and this is now reflected in the graphs/texts.  @ebezzam , @bmcfee let me know if there are additional points or the we can move on ;).
