### 2023-06-22 JorenSix/Olaf#35 @ebezzam
Got it, could you still format the citation for "Cortes et. al, 2022" entry so that it's like the others? The long list of names on lines 41-42 and 51-52 is a but awkward.
