### 2023-06-22 #5459 @ebezzam
@JorenSix thanks for quickly addressing the points! Paper is looking much more complete, and great that you could resolve this bottleneck! Just one more formatting :see_no_evil: https://github.com/JorenSix/Olaf/issues/35
