+++ paper.bib
@@ -53,7 +53,7 @@
 
 @article{holighaus2012framework,
   title={A framework for invertible, real-time constant-Q transforms},
-  author={{Holighaus, Nicki and D{\"o}rfler, Monika and Velasco, Gino Angelo and Grill, Thomas}},
+  author={Holighaus, Nicki and D{\"o}rfler, Monika and Velasco, Gino Angelo and Grill, Thomas},
   journal={IEEE Transactions on Audio, Speech, and Language Processing},
   volume={21},
   number={4},
@@ -63,7 +63,7 @@
   doi={10.1109/TASL.2012.2234114}
 }
 
-@InProceedings{six2018dupapps,
+@inproceedings{six2018dupapps,
   author="Six, Joren and Bressan, Federica and Leman, Marc",
   editor="Serra, Giuseppe and Tasso, Carlo",
   title="Applications of Duplicate Detection in Music Archives: 
@@ -135,7 +135,7 @@
   publisher={ACM New York, NY, USA}
 }
 
-@INPROCEEDINGS{9044274,
+@inproceedings{9044274,
   author={Mehmood, Zahid and Ashfaq Qazi, Khurram and Tahir, Muhammad and Muhammad Yousaf, Rehan and Sardaraz, Muhammad},
   booktitle={6th Conference on Data Science and Machine Learning Applications (CDMA 2020)},
   title={Potential Barriers to Music Fingerprinting Algorithms in the Presence of Background Noise}, 
@@ -145,7 +145,7 @@
   pages={25-30},
   doi={10.1109/CDMA47397.2020.00010}}
   
-  @inproceedings{dlfm2017_mirchiving,
+@inproceedings{dlfm2017_mirchiving,
  author = {de Valk, Reinier and Volk, Anja and Holzapfel, Andre and 
            Pikrakis, Aggelos and Kroher, Nadine and Six, Joren},
  title = {MIRchiving: Challenges and Opportunities of Connecting 
@@ -212,7 +212,7 @@
 }
 
 @inproceedings{six2021panakovtwo,
-  author       = {{Six, Joren}},
+  author       = {Six, Joren},
   booktitle    = {{Extended abstracts for the Late-Breaking Demo Session of the 22st International Society for Music Information Conference (ISMIR 2021) }},
   keywords     = {{Acoustic fingerprinting}},
   language     = {{eng}},
@@ -337,13 +337,13 @@
 
 
 @inproceedings{cortes2022baf,
-  author       = {{Cortès, Guillem and Ciurana, Alex and Molina, Emilio and Miron, Marius and Meyers, Owen and Six, Joren and Serra, Xavier}},
+  author       = {Cortès, Guillem and Ciurana, Alex and Molina, Emilio and Miron, Marius and Meyers, Owen and Six, Joren and Serra, Xavier},
   booktitle    = {{Proceedings of the 23rd International Society for Music Information Retrieval Conference (ISMIR 2022)}},
   keywords     = {{Acoustic fingerprinting}},
   language     = {{eng}},
   location     = {{Bengalaru, India}},
   pages        = {{908--916}},
-  title        = {{BAF : an audio fingerprinting dataset for broadcast monitoring}},
+  title        = {{BAF: an audio fingerprinting dataset for broadcast monitoring}},
   url          = {{http://dx.doi.org/10.5281/zenodo.7343030}},
   doi          = {{10.5281/zenodo.7343030}},
   year         = {{2022}},
