### 2023-06-23 JorenSix/Olaf#35 @JorenSix
Fixed, the problem was the use of `authors = {{author list}}`. With this formatting the author list is used literally, when using `authors = {author list}` it is correctly replaced by an 'et al'.
