### 2023-06-23 #5459 @JorenSix
The formatting of the author names is fixed, the use of `authors={{auhor list}}` made latex quote the long author list literally. The use of `authors={author list}` fixed it.

Thanks again @ebezzam  for spotting the issue!
