### 2023-06-23 6e32fb8
Performance upgrade by using ARM NEON SIMD version for macos of max filter in time
