+++ src/olaf_fp_ref_mem.h
@@ -5,1887 +5,2080 @@
 
 /**The array with fingerprints.*/
 const uint64_t olaf_db_mem_fps[] = {
-5118305822,
-5118306302,
-5118306782,
-5118307262,
-5118372961,
-5118492826,
-5118493306,
-5118559802,
-5118561722,
-5118562013,
-5118562201,
-5118562681,
-5118563161,
-5118563641,
-5118564122,
-5118565562,
-5118565853,
-5118566041,
-5118566521,
-5118567001,
-5118567481,
-5118567481,
-5118627617,
-5118628097,
-5118629057,
-5118631457,
-5118631937,
-5118632417,
-5118632897,
-5118699394,
-5118699874,
-5118700354,
-5118757729,
-5118758209,
-5118760569,
-5118760609,
-5118761089,
-5118761569,
-5118762049,
-5118886282,
-5118886762,
-5118887242,
-5118887722,
-5118888202,
-5118888682,
-5118889162,
-5118892042,
-5118892522,
-5118893002,
-5119021704,
-6174938771,
-6175002626,
-6175003106,
-6175003586,
-6175006466,
-6175007426,
-6175007906,
-6175328187,
-6175328667,
-6175329147,
-6175332027,
-6175332507,
-6175332987,
-6175333467,
-6175335867,
-6175336347,
-6175336827,
-6175395281,
-6175397681,
-6175399121,
-6175401521,
-6175402001,
-6175402961,
-6175457456,
-6175457936,
-6175461657,
-6175462256,
-6175463942,
-6175465977,
-6175466457,
-6175467782,
-6175526832,
-6175527312,
-6175528272,
-6175530672,
-6175531152,
-6175531632,
-6175532112,
-6175533237,
-6175534277,
-6175657785,
-6175658745,
-6175662105,
-6175988749,
-6175991629,
-6175992589,
-6176114982,
-6176115462,
-6176115942,
-14781583317,
-14781587157,
-14781587637,
-14781588117,
-19059902414,
-23355917670,
-23372569185,
-117861781224,
-117861785064,
-119991572100,
-119991635716,
-119991639556,
-129656493639,
-194097195082,
-194097195563,
-194097261578,
-203760801546,
-204834474917,
-237046268027,
-238119944796,
-238119945756,
-241341180122,
-279996661793,
-279996989654,
-279996990134,
-279996991094,
-279996991574,
-281069750086,
-282127565045,
-282127565525,
-282127566485,
-282127566965,
-283200456714,
-284274198058,
-284274458683,
-284274459163,
-284274459643,
-284274460123,
-287513120485,
-293938731317,
-297160016055,
-297160016535,
-299323359294,
-299323359774,
-311134462284,
-311135375948,
-313265168731,
-313266213481,
-348716468818,
-348716469298,
-348716472178,
-348716472658,
-348716473138,
-408844960782,
-408844961262,
-412050391433,
-412050391913,
-412050392393,
-412050392873,
-412050397193,
-412050397673,
-413140784175,
-413140848751,
-418508645673,
-468975036011,
-468975619814,
-468975623654,
-479712776097,
-479712781497,
-518349851339,
-560225459735,
-560225846046,
-560225846526,
-560226042174,
-562373267173,
-564521082371,
-565594816810,
-565594817770,
-574184423742,
-574184424222,
-574184424702,
-574184425662,
-574184426142,
-574184426622,
-574184427102,
-574184429502,
-574184430942,
-574184489982,
-574184490462,
-574184490942,
-574184491902,
-574184492382,
-574184494782,
-574184495262,
-574184495742,
-574184496222,
-574184498622,
-574184499102,
-574184499582,
-574184500062,
-574184556957,
-574185472062,
-574185472542,
[…]
