+++ src/olaf_ep_extractor.c
@@ -33,10 +33,6 @@ struct Olaf_EP_Extractor{
 	// The 'vertical' max-filtered magnitudes for the  
 	float** maxes;
 
-	// The horizontally max-filtered magnitudes for the
-	// current frame
-	float* horizontalMaxes;
-
 	int filterIndex;
 	
 	int audioBlockIndex;
@@ -54,12 +50,9 @@ Olaf_EP_Extractor * olaf_ep_extractor_new(Olaf_Config * config){
 
 	int halfAudioBlockSize = config->audioBlockSize / 2;
 
-	ep_extractor->horizontalMaxes = (float *) calloc(halfAudioBlockSize ,sizeof(float));
-  if(ep_extractor->horizontalMaxes == NULL) fprintf(stdout,"Failed to allocate memory: horizontalMaxes");
-
 	ep_extractor->eventPoints.eventPoints = (struct eventpoint *) calloc(config->maxEventPoints , sizeof(struct eventpoint));
 	ep_extractor->eventPoints.eventPointIndex = 0;
-  if(ep_extractor->horizontalMaxes == NULL) fprintf(stdout,"Failed to allocate memory: eventPoints");
+	if(ep_extractor->eventPoints.eventPoints == NULL) fprintf(stdout,"Failed to allocate memory: eventPoints");
 
 	//initialize t with a high number
 	for(int i = 0 ; i < config->maxEventPoints; i++){
@@ -67,44 +60,64 @@ Olaf_EP_Extractor * olaf_ep_extractor_new(Olaf_Config * config){
 	}
 	
 	ep_extractor->mags  =(float **) calloc(config->filterSizeTime , sizeof(float *));
-  if(ep_extractor->mags == NULL) fprintf(stdout,"Failed to allocate memory: mags");
+	if(ep_extractor->mags == NULL) fprintf(stdout,"Failed to allocate memory: mags");
 	ep_extractor->maxes = (float **) calloc(config->filterSizeTime , sizeof(float *));
-  if(ep_extractor->maxes == NULL) fprintf(stdout,"Failed to allocate memory: maxes");
+	if(ep_extractor->maxes == NULL) fprintf(stdout,"Failed to allocate memory: maxes");
 
 	for(int i = 0; i < config->filterSizeTime ;i++){
 		ep_extractor->maxes[i]= (float *) calloc(halfAudioBlockSize , sizeof(float));
-    if(ep_extractor->maxes[i] == NULL) fprintf(stdout,"Failed to allocate memory: maxes[i]");
+		if(ep_extractor->maxes[i] == NULL) fprintf(stdout,"Failed to allocate memory: maxes[i]");
+
 		ep_extractor->mags[i] = (float *) calloc(halfAudioBlockSize , sizeof(float));
-    if(ep_extractor->mags[i] == NULL) fprintf(stderr,"Failed to allocate memory: mags[i]");
+		if(ep_extractor->mags[i] == NULL) fprintf(stderr,"Failed to allocate memory: mags[i]");
 	}
 
-  	ep_extractor->filterIndex = 0;
-  	return ep_extractor;
+	ep_extractor->filterIndex = 0;
+	return ep_extractor;
 }
 
 void olaf_ep_extractor_destroy(Olaf_EP_Extractor * ep_extractor){
 	free(ep_extractor->eventPoints.eventPoints);
-	free(ep_extractor->horizontalMaxes);
 
 	for(int i = 0; i < ep_extractor->config->filterSizeTime ;i++){
 	  free(ep_extractor->maxes[i]);
 	  free(ep_extractor->mags[i]);
-  	}
-
-  	free(ep_extractor->maxes);
-  	free(ep_extractor->mags);
[…]
