+++ eval/olaf_vs_panako/olaf_vs_panako_plot_query.R
@@ -14,7 +14,7 @@ a <- ggplot(bm, aes(bm$index_size,bm$olaf_query_time)) +
   geom_point(na.rm = TRUE) +
   xlab("Indexed audio (s)") +
   ylab("Query time (s/s)") +
-  ylim(c(0,300))+
+  ylim(c(0,500))+
   scale_color_manual(name='Query times',
                      breaks=c('Olaf', 'Panako'),
                      values=c('Olaf'='black', 'Panako'='steelblue'))+
