+++ Makefile
@@ -23,6 +23,29 @@ compile:
 	mkdir -p bin
 	gcc -o bin/olaf_c *.o 			-lc -lm -ffast-math -pthread
 
+lib:
+	gcc -c src/pffft.c 					-W -Wall -std=gnu11 -pedantic -O2 #pfft needs M_PI and other constants not in the ANSI c standard
+	gcc -c src/midl.c 					-W -Wall -std=c11 -pedantic -O2
+	gcc -c src/mdb.c 					-W -Wall -std=c11 -pedantic -O2
+	gcc -c src/hash-table.c     		-W -Wall -std=c11 -pedantic -O2
+	gcc -c src/queue.c  		   		-W -Wall -std=c11 -pedantic -O2
+	gcc -c src/olaf_deque.c  	   		-W -Wall -std=c11 -pedantic -O2
+	gcc -c src/olaf_max_filter_perceptual_van_herk.c	-W -Wall -std=c11 -pedantic -O2
+	gcc -c src/olaf.c 					-W -Wall -std=c11 -pedantic -O2
+	gcc -c src/olaf_fp_file_writer.c 	-W -Wall -std=c11 -pedantic -O2
+	gcc -c src/olaf_db.c 				-W -Wall -std=c11 -pedantic -O2
+	gcc -c src/olaf_fp_db_writer.c 		-W -Wall -std=c11 -pedantic -O2
+	gcc -c src/olaf_fp_db_writer_cache.c -W -Wall -std=c11 -pedantic -O2
+	gcc -c src/olaf_ep_extractor.c 		-W -Wall -std=c11 -pedantic -O2
+	gcc -c src/olaf_fp_extractor.c 		-W -Wall -std=c11 -pedantic -O2
+	gcc -c src/olaf_reader_stream.c 	-W -Wall -std=c11 -pedantic -O2
+	gcc -c src/olaf_runner.c 			-W -Wall -std=c11 -pedantic -O2
+	gcc -c src/olaf_stream_processor.c 	-W -Wall -std=c11 -pedantic -O2
+	gcc -c src/olaf_fp_matcher.c 		-W -Wall -std=c11 -pedantic -O2
+	gcc -c src/olaf_config.c 			-W -Wall -std=c11 -pedantic -O2
+	mkdir -p bin
+	gcc -o bin/olaf.so *.o 			-lc -lm -ffast-math -pthread -shared
+
 #A compilation with support for profiling
 compile_gprof:
 	gcc -c src/pffft.c 					-pg -W -Wall -std=gnu11 -pedantic -O2 #pfft needs M_PI and other constants not in the ANSI c standard
@@ -152,7 +175,6 @@ test:
 	gcc -o bin/olaf_tests *.o		-lc -lm -ffast-math
 	mkdir -p tests/olaf_test_db
 	- rm tests/olaf_test_db/*
-
 
 #Generate doxygen API documentation
 docs:
