### 2023-06-29 #5459 @bmcfee
I read through the paper, and have a few comments:

- Line 38: "MIR" used without definition
- Line 43: "similarly specced" could be more precisely worded
- Line 72: "file handling and transcoding is" → "... are"

General comment: in the bibliography entries, many parts that should be capitalized are not, e.g., "Lmdb".  You can force proper capitalization by embedding the appropriate parts in curlies, eg, `"{LMDB}: A general-purpose ..."` .

Citation for Lòpez-García et al., seems to have an awkwardly formatted DOI (ie it's URL-encoded).  Please check this one.
