+++ paper.bib
@@ -389,7 +389,7 @@
 
 @misc{pffft,
   author = {Julien Pommier},
-  title = {pffft:  a pretty fast FFT},
+  title = {{PFFFT}: a pretty fast {FFT}},
   year = {2022},
   publisher = {BitBucket},
   journal = {BitBucket repository},
@@ -407,7 +407,7 @@
 
 @misc{lmdb,
   author = {Howard Chu},
-  title = {lmdb: a general-purpose programming language and toolchain for maintaining robust, optimal, and reusable software.},
+  title = {{LMDB}: a general-purpose programming language and toolchain for maintaining robust, optimal, and reusable software.},
   year = {2022},
   publisher = {OpenLDAP},
   journal = {OpenLDAP repository},
@@ -457,6 +457,6 @@ year="2022",
 publisher="Springer International Publishing",
 address="Cham",
 pages="192--204",
-doi          = {{10.1007/978-3-031-07015-0_16}},
+doi          = "10.1007/978-3-031-07015-0_16",
 }
 
