+++ paper.bib
@@ -2,11 +2,11 @@
 @inproceedings{six2012ethnic_fingerprinter,
   author = {Joren Six and Olmo Cornelis},
   booktitle = {{Proceedings of the Folk Music Analysis conference (FMA 2012)}},
-  title = {{A Robust Audio Fingerprinter Based on Pitch Class Histograms - Applications for Ethnic Music Archives}},
+  title = {A Robust Audio Fingerprinter Based on Pitch Class Histograms - Applications for Ethnic Music Archives},
   year = {2012}}
 
 @inproceedings{velasco2011constructing,
-  title={Constructing an invertible constant-Q transform with non-stationary Gabor frames},
+  title={Constructing an invertible constant-{Q} transform with non-stationary Gabor frames},
   author={{Velasco, Gino Angelo and Holighaus, Nicki and D{\"o}rfler, Monika and Grill, Thomas}},
   booktitle={{Proceedings of the 14th International Conference on Digital Audio Effects (DAFx-11)}},
   volume={33},
@@ -23,7 +23,7 @@
 
 
 @inproceedings{herre2002scalable,
-  title={Scalable robust audio fingerprinting using MPEG-7 content description},
+  title={Scalable robust audio fingerprinting using {MPEG-7} content description},
   author={Herre, J{\"u}rgen and Hellmuth, Oliver and Cremer, Markus},
   booktitle={Multimedia Signal Processing, 2002 IEEE Workshop on},
   pages={165--168},
@@ -37,7 +37,7 @@
   booktitle = {Proceedings of the 4th {I}nternational {S}ymposium on {M}usic {I}nformation {R}etrieval ({ISMIR} 2003)},
   location  = {Baltimore, MD, USA},
   pages     = {7--13},
-  title     = {{An Industrial-Strength Audio Search Algorithm}},
+  title     = {An Industrial-Strength Audio Search Algorithm},
   year      = {2003},
   doi       = {10.5281/zenodo.1416340}
 }
@@ -169,12 +169,12 @@
 @inproceedings{sonnleitner2014quad_based_fingerprinter,
   author    = {Reinhard Sonnleitner and Gerhard Widmer},
   booktitle = {{Proceedings of the 17th International Conference on Digital Audio Effects (DAFx-14)}},
-  title     = {{Quad-based Audio Fingerprinting Robust To Time And Frequency Scaling}},
+  title     = {{Quad-based audio fingerprinting robust to time and frequency scaling}},
   year      = {2014},
 }
 
 @inproceedings{ellis2014labrosafp,
-  title={The 2014 labrosa audio fingerprint system},
+  title={The 2014 {LabROSA} audio fingerprint system},
   author={Ellis, Daniel},
   booktitle={{MIREX abstracts of the 15th International Symposium on Music Information Retrieval (ISMIR 2014)}},
   year={2014}
@@ -398,7 +398,7 @@
 
 @misc{zig,
   author = {Andrew Kelley and contributors},
-  title = {zig: a general-purpose programming language and toolchain for maintaining robust, optimal, and reusable software.},
+  title = {Zig: a general-purpose programming language and toolchain for maintaining robust, optimal, and reusable software.},
   year = {2023},
   publisher = {GitHub},
   journal = {GitHub repository},
