+++ paper.bib
@@ -345,7 +345,7 @@
   pages        = {{908--916}},
   title        = {{BAF: an audio fingerprinting dataset for broadcast monitoring}},
   url          = {{http://dx.doi.org/10.5281/zenodo.7343030}},
-  doi          = {{10.5281/zenodo.7343030}},
+  doi          = {10.5281/zenodo.7343030},
   year         = {{2022}},
 }
 
