### 2023-06-29 #5459 @JorenSix
Fixed, I did miss that doi. FYI again the {{}} vs {} problem. Not sure where the url encoding kicks in but it is fine now.
