### 2023-06-29 #5459 @bmcfee
@JorenSix looks like you have a duplicate citation key in your bib file and it's breaking the paper build:

https://github.com/JorenSix/Olaf/blob/d474bf6bbd6ad8d6c2562f81415b6316b2cfea77/paper.bib#L233

https://github.com/JorenSix/Olaf/blob/d474bf6bbd6ad8d6c2562f81415b6316b2cfea77/paper.bib#L317
