+++ paper.bib
@@ -230,17 +230,7 @@
   doi           = {10.5281/zenodo.1416190}
 }
 
-@article{six2015synchronizing,
-  title={Synchronizing multimodal recordings using audio-to-audio alignment},
-  author={Six, Joren and Leman, Marc},
-  journal={Journal on Multimodal User Interfaces},
-  volume={9},
-  number={3},
-  pages={223--229},
-  year={2015},
-  publisher={Springer},
-  doi="10.1007/s12193-015-0196-1"
-}
+
 
 @inproceedings{six2017framework,
   title={A framework to provide fine-grained time-dependent context for active listening experiences},
