### 2023-07-03 #5459 @danielskatz
@JorenSix - I've suggested some minor changes in https://github.com/JorenSix/Olaf/pull/39. Please merge this, or let me know what you disagree with.

Also in the paper, I see "C also provides a lot of exiting footguns, of which I made ample use." Since this is a term that is not in wide use, if you want to keep it, I suggest explaining it. I also wonder if this sentence should just be deleted, since it doesn't contribute to the text in that paragraph in a way that makes sense to me.
