### 2023-07-03 #5459 @JorenSix
Merged, thanks for the fixes.
Now there is a new version of the paper with a clearer sentence: Developing bug-free code in the C programming language is notoriously challenging. However, many bugs have been found by ...
