### 2023-07-03 #5459 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
