### 2023-07-03 #5459 @danielskatz
Congratulations to @JorenSix on your publication!!

And thanks to @liscio and @ebezzam for reviewing, and to @bmcfee for editing!!
JOSS depends on volunteers and couldn't do this without you!
