### 2023-07-03 #5459 @JorenSix
Great!
Thanks you @liscio, @ebezzam and @bmcfee for taking the time and for the comments. The review process certainly improved the paper and the Olaf software itself! I do think the JOSS concept is working as it should. Thanks again!
